refactor: split the seeds page and fix the descriptions
